Move `CreateComment` command
